area math booleans working with bugs improved
